A lot of implementation in FAQ.html and Reqommendations.html
FAQ.html This file it's finished. A lot of questions about cloud been added along with their answers.

Reqommendations.html In this file a form it's added so the user can fill in the informations about the user and what informations he/she wants to add to the cloud. More options will be added soon.

NOW Terms and Conditions PDF been added TO ALL html files. User now can see the terms by going to footer and clicking on Term text